+++ elide-annotations/src/main/java/com/yahoo/elide/annotation/OnCreatePreSecurity.java
@@ -19,4 +19,13 @@ import java.lang.annotation.Target;
 @Target({ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface OnCreatePreSecurity {
+    /**
+     * Field name on which the annotated method is only triggered if that field is modified.
+     * If value is empty string, then trigger once when the object is created.
+     * If value is "*", then this method will be triggered once for each field that
+     * the user sent in the creation request.
+     *
+     * @return the field name that triggers the method
+     */
+    String value() default "";
 }