+++ elide-annotations/src/main/java/com/yahoo/elide/annotation/OnReadPreCommit.java
@@ -23,8 +23,9 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface OnReadPreCommit {
     /**
-     * Field name on which the annotated method is only triggered if that field is modified.
-     * If value is empty string, then trigger for any modification of the object.
+     * Field name on which the annotated method is only triggered if that field is read.
+     * If value is empty string, then trigger once when the object is read.
+     * If value is "*", then trigger for all field reads.
      *
      * @return the field name that triggers the method
      */