+++ elide-annotations/src/main/java/com/yahoo/elide/annotation/OnUpdatePreSecurity.java
@@ -22,7 +22,8 @@ import java.lang.annotation.Target;
 public @interface OnUpdatePreSecurity {
     /**
      * Field name on which the annotated method is only triggered if that field is modified.
-     * If value is empty string, then trigger for any modification of the object.
+     * If value is empty string, then trigger once when the object is updated.
+     * If value is "*", then trigger for all field modifications.
      *
      * @return the field name that triggers the method
      */