+++ elide-core/src/main/java/com/yahoo/elide/core/EntityDictionary.java
@@ -803,6 +803,11 @@ public class EntityDictionary {
         return getEntityBinding(cls).getTriggers(annotationClass, fieldName);
     }
 
+    public <A extends Annotation> Collection<LifeCycleHook> getTriggers(Class<?> cls,
+                                                                        Class<A> annotationClass) {
+        return getEntityBinding(cls).getTriggers(annotationClass);
+    }
+
     /**
      * Return a single annotation from field or accessor method.
      *
@@ -1000,7 +1005,8 @@ public class EntityDictionary {
     }
 
     /**
-     * Binds a lifecycle hook to a particular field or method in an entity.
+     * Binds a lifecycle hook to a particular field or method in an entity.  The hook will be called a
+     * single time per request per field READ, CREATE, or UPDATE.
      * @param entityClass The entity that triggers the lifecycle hook.
      * @param annotationClass (OnReadPostCommit, OnUpdatePreSecurity, etc)
      * @param fieldOrMethodName The name of the field or method
@@ -1014,7 +1020,31 @@ public class EntityDictionary {
     }
 
     /**
-     * Binds a lifecycle hook to a particular entity class.
+     * Binds a lifecycle hook to a particular entity class.  The hook will either be called:
+     *  - A single time single time per request per class READ, CREATE, UPDATE, or DELETE.
+     *  - Multiple times per request per field READ, CREATE, or UPDATE.
+     *
+     * The behavior is determined by the value of the {@code allowMultipleInvocations} flag.
+     * @param entityClass The entity that triggers the lifecycle hook.
+     * @param annotationClass (OnReadPostCommit, OnUpdatePreSecurity, etc)
+     * @param callback The callback function to invoke.
+     * @param allowMultipleInvocations Should the same life cycle hook be invoked multiple times for multiple
+     *                              CRUD actions on the same model.
+     */
+    public void bindTrigger(Class<?> entityClass,
+                            Class<? extends Annotation> annotationClass,
+                            LifeCycleHook callback,
+                            boolean allowMultipleInvocations) {
+        if (allowMultipleInvocations) {
+            getEntityBinding(entityClass).bindTrigger(annotationClass, callback);
+        } else {
+            getEntityBinding(entityClass).bindTrigger(annotationClass, PersistentResource.CLASS_NO_FIELD, callback);
+        }
+    }
+
+    /**
+     * Binds a lifecycle hook to a particular entity class.   The hook will be called a single time per request
+     * per class READ, CREATE, UPDATE, or DELETE.
      * @param entityClass The entity that triggers the lifecycle hook.
      * @param annotationClass (OnReadPostCommit, OnUpdatePreSecurity, etc)
      * @param callback The callback function to invoke.
@@ -1022,8 +1052,9 @@ public class EntityDictionary {
     public void bindTrigger(Class<?> entityClass,
                             Class<? extends Annotation> annotationClass,
                             LifeCycleHook callback) {
-        getEntityBinding(entityClass).bindTrigger(annotationClass, PersistentResource.CLASS_NO_FIELD, callback);
+        bindTrigger(entityClass, annotationClass, callback, false);
     }
+
 
     /**
      * Returns true if the relationship cascades deletes and false otherwise.